Prevent TDE SMGR from engaging with tde_heap_basic
1. Currently, tde smgr functions (like tde_mdwritev) will encrypt/decrypt
tde_heap_basic relations because the only rule of engagement is the
existence of the internal key. And any tde_heap_basic would have one.
This commit creates a rel_type for the internal key so tde smgr can
distinguish if that key (hence relation) was meant for the "tde_heap".

2. During ALTER ... TABLESPACE, tde_mdcreate gets called and
tde_smgr_get_key is trying to get the internal key. But there is no key
moved to the new table space since it is called during
RelationCreateStorage() and pg_tde.map file might not exist. And it's ok in
this case but it results in an error. This commit introduces an option not
to have a pg_tde.map (treated as no key).